requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via unstructured-client
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.0
    # via langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
backoff==2.2.1
    # via unstructured
beautifulsoup4==4.13.4
    # via unstructured
boto3==1.38.23
    # via -r requirements.in
botocore==1.38.23
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   pinecone
    #   requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via unstructured
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   python-oxmsg
    #   uvicorn
cryptography==45.0.2
    # via unstructured-client
dataclasses-json==0.6.7
    # via
    #   langchain-community
    #   unstructured
emoji==2.14.1
    # via unstructured
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.115.12
    # via -r requirements.in
filetype==1.2.0
    # via unstructured
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
html5lib==1.1
    # via unstructured
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   langsmith
    #   unstructured-client
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.1
    # via nltk
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.24
    # via -r requirements.in
langchain-core==0.3.61
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.3.8
    # via
    #   -r requirements.in
    #   langchain
langdetect==1.0.9
    # via unstructured
langsmith==0.3.42
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==5.4.0
    # via
    #   python-docx
    #   unstructured
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via unstructured-client
nltk==3.9.1
    # via unstructured
numpy==2.2.6
    # via
    #   langchain-community
    #   unstructured
olefile==0.47
    # via python-oxmsg
openpyxl==3.1.5
    # via -r requirements.in
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pytest
pinecone==7.0.1
    # via -r requirements.in
pinecone-plugin-interface==0.0.7
    # via pinecone
pluggy==1.6.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via unstructured
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
    #   unstructured-client
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via
    #   -r requirements.in
    #   langchain-community
pypdf==5.5.0
    # via
    #   -r requirements.in
    #   unstructured-client
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.26.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pinecone
python-docx==1.1.2
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-iso639==2025.2.18
    # via unstructured
python-magic==0.4.27
    # via unstructured
python-oxmsg==0.0.2
    # via unstructured
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
rapidfuzz==3.13.0
    # via unstructured
regex==2024.11.6
    # via
    #   nltk
    #   tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
    #   unstructured
requests-toolbelt==1.0.0
    # via
    #   langsmith
    #   unstructured-client
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via
    #   html5lib
    #   langdetect
    #   python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
starlette==0.46.2
    # via fastapi
structlog==25.3.0
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.9.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   nltk
    #   unstructured
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   langchain-core
    #   pinecone
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-oxmsg
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
    #   unstructured
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
unstructured==0.17.2
    # via -r requirements.in
unstructured-client==0.35.0
    # via unstructured
urllib3==2.4.0
    # via
    #   botocore
    #   pinecone
    #   requests
uvicorn[standard]==0.34.2
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
webencodings==0.5.1
    # via html5lib
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via unstructured
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
